requirements.txt
PIL
eel
google-generativeai
